requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

python-telegram-bot[ext, job-queue]>=21.0.1,<22.0.0
httpx>=0.25.0,<1.0.0

google-cloud-secret-manager>=2.16.0
google-api-python-client>=2.80.0
oauth2client>=4.1.3
gspread>=5.10.0
google-generativeai>=0.8.0

python-dotenv>=1.0.0
dateparser>=1.1.0

# Optional: Add any other specific dependencies from your project
# e.g., pandas if used, specific DB drivers, etc. 